Added authentication redirects on a page level before any queries occur for retrieving a specific submission or a specific user's submission.
