+++ js/src/app/submission/s/[submissionId]/component/SubmissionDetailsContent.module.css
