+++ js/src/lib/router.tsx
@@ -1,11 +1,11 @@
-import SubmissionDetails from "@/app/submission/s/[submissionId]/SubmissionDetails.page";
 import DashboardPage from "@/app/dashboard/Dashboard.page";
 import LeaderboardPage from "@/app/leaderboard/Leaderboard.page";
 import LoginPage from "@/app/login/Login.page";
 import Onboarding from "@/app/onboarding/Onboarding.page";
 import RootPage from "@/app/Root.page";
+import SubmissionDetailsPage from "@/app/submission/s/[submissionId]/SubmissionDetails.page";
+import UserSubmissionsPage from "@/app/submission/u/[userId]/UserSubmissions.page";
 import { createBrowserRouter } from "react-router-dom";
-import UserSubmissions from "@/app/submission/u/[userId]/UserSubmissions.page";
 
 export const router = createBrowserRouter([
   {
@@ -30,10 +30,10 @@ export const router = createBrowserRouter([
   },
   {
     path: "/submission/s/:submissionId",
-    element: <SubmissionDetails />,
+    element: <SubmissionDetailsPage />,
   },
   {
     path: "/submission/u/:userId",
-    element: <UserSubmissions />,
+    element: <UserSubmissionsPage />,
   },
 ]);
